ASC-1418 Add scenario option
This commit adds the `--scenario` option to moleculerize and sets its
default to `default`.  This value is used in the `molecule.yml.j2`
template to define the scenario used in the final `molecule.yml` file.
It is left up to the user to define the desired output file which
retains the existing behavior of this tool.